requirements.txt
requests
beautifulsoup4
pandas
numpy==1.24.3
psycopg2-binary==2.9.10
PyYAML==6.0